Add FREEZE_SERVICE_SECONDARY_INSTANCE_ON_START rc parameter
Upon reboot of all cluster nodes, the services will all
start on the first node starting openha. If this behaviour
is not acceptable in your context, this option
can be set to "yes" to force the start of only the
services the node is primary of. The service instances
the node is secondary of are transitioned to the
FROZEN_STOP state, and thus the unfreeze command will
be necessary to restore HA after all nodes are booted.